+++ README.md
@@ -30,12 +30,12 @@ Or install it yourself as:
 It's super easy to get started. In your Ruby shell ($ irb), for example, do this:
 
     require 'block_io'
-    BlockIo.set_options :api_key => 'API KEY', :pin => 'SECRET PIN', :version => 2
+    block_io = BlockIo::Client.new api_key: 'API KEY', pin: 'SECRET PIN'
      
 And you're good to go:
 
-    BlockIo.get_new_address
-    BlockIo.get_my_addresses
+    block_io.get_new_address
+    block_io.get_my_addresses
 
 For more information, see https://block.io/api/simple/ruby
