+++ examples/basic.rb
@@ -7,22 +7,24 @@
 
 require 'block_io'
 
-puts BlockIo.set_options :api_key => ENV['API_KEY'], :pin => ENV['PIN'], :version => 2
+block_io = BlockIo::Client.new api_key: ENV['API_KEY'], pin: ENV['PIN']
+puts block_io.get_balance
+puts block_io.get_my_addresses
 
 begin
-  puts BlockIo.get_new_address(:label => 'testDest')
+  puts block_io.get_new_address(:label => 'testDest')
 rescue Exception => e
   # if this failed, we probably created testDest label before
   puts e.to_s
 end
 
-puts BlockIo.withdraw_from_labels(:from_labels => 'default', :to_label => 'testDest', :amount => '2.5')
+puts block_io.withdraw_from_labels(:from_labels => 'default', :to_label => 'testDest', :amount => '2.5')
 
-puts BlockIo.get_address_balance(:labels => 'default,testDest')
+puts block_io.get_address_balance(:labels => 'default,testDest')
 
-puts BlockIo.get_transactions(:type => 'sent') # API v2 only
+puts block_io.get_transactions(:type => 'sent') # API v2 only
 
-puts BlockIo.get_transactions(:type => 'received') # API v2 only
+puts block_io.get_transactions(:type => 'received') # API v2 only
 
-puts BlockIo.get_current_price(:base_price => 'BTC')
+puts block_io.get_current_price(:base_price => 'BTC')
